+++ node/event-to-sentry.js
@@ -1,7 +1,7 @@
 const sqlite3 = require('sqlite3').verbose();
 const Sentry = require('@sentry/node');
 
-// Put your DSN key here
+// TODO Put your DSN key here
 const DSN = 'http://211dbbd50f41437a83316cdd4bec7513@localhost:9000/4'
 
 var eventSavedOffline;
@@ -9,11 +9,13 @@ var eventSavedOffline;
 Sentry.init({ 
     dsn: DSN,
     beforeSend: function (event,hint) {
-        // Normally you return the 'event' here, but in our case we're returning the event that we loaded from sqlite (i.e. it was saved offline)
+        // We're re-purposing beforeSend to send our event loaded from offline database, as opposed to the event generated below by Sentry.captureException('ignore me')
         return eventSavedOffline
+        // `return event` or `return null`
     }
 });
 
+// TODO put Node module and path for DB you'll write to
 let db = new sqlite3.Database('./sqlite.db', sqlite3.OPEN_READWRITE, (err) => {
     if (err) {
       return console.error(err.message);
@@ -27,8 +29,8 @@ db.serialize(() => {
         console.error(err.message);
       }
       
-      // the 5th event in my database is a Node Event. All the rest are Python Events, at the moment.
-      let row = rows[4]
+      // Selecting the most recent event I wrote to the database
+      let row = rows[7]
       let string = new Buffer(row.data).toString('ascii')
       eventSavedOffline = JSON.parse(string)
